+++ packages/plugin-gallery/web/src/icons/ColumnsIconNew.tsx
@@ -1,0 +1,24 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const ColumnsIconNew = props => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M6.5 19H9.5C10.3 19 11 18.3 11 17.5V6.5C11 5.7 10.3 5 9.5 5H6.5C5.7 5 5 5.7 5 6.5V17.5C5 18.3 5.7 19 6.5 19ZM6 6.5C6 6.2 6.2 6 6.5 6H9.5C9.8 6 10 6.2 10 6.5V17.5C10 17.8 9.8 18 9.5 18H6.5C6.2 18 6 17.8 6 17.5V6.5ZM14.5 19H17.5C18.3 19 19 18.3 19 17.5V6.5C19 5.7 18.3 5 17.5 5H14.5C13.7 5 13 5.7 13 6.5V17.5C13 18.3 13.7 19 14.5 19ZM14 6.5C14 6.2 14.2 6 14.5 6H17.5C17.8 6 18 6.2 18 6.5V17.5C18 17.8 17.8 18 17.5 18H14.5C14.2 18 14 17.8 14 17.5V6.5Z"
+        fill="black"
+      />
+    </svg>
+  );
+};
+
+export default ColumnsIconNew;