+++ packages/plugin-gallery/web/src/icons/PanoramaIconNew.tsx
@@ -1,0 +1,24 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const PanoramaIconNew = props => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M6.5 11H17.5C18.3 11 19 10.3 19 9.5V6.5C19 5.7 18.3 5 17.5 5H6.5C5.7 5 5 5.7 5 6.5V9.5C5 10.3 5.7 11 6.5 11ZM6 6.5C6 6.2 6.2 6 6.5 6H17.5C17.8 6 18 6.2 18 6.5V9.5C18 9.8 17.8 10 17.5 10H6.5C6.2 10 6 9.8 6 9.5V6.5ZM6.5 19H17.5C18.3 19 19 18.3 19 17.5V14.5C19 13.7 18.3 13 17.5 13H6.5C5.7 13 5 13.7 5 14.5V17.5C5 18.3 5.7 19 6.5 19ZM6 14.5C6 14.2 6.2 14 6.5 14H17.5C17.8 14 18 14.2 18 14.5V17.5C18 17.8 17.8 18 17.5 18H6.5C6.2 18 6 17.8 6 17.5V14.5Z"
+        fill="black"
+      />
+    </svg>
+  );
+};
+
+export default PanoramaIconNew;