+++ packages/plugin-gallery/web/src/icons/SlidesIconNew.tsx
@@ -1,0 +1,22 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const SlidesIconNew = props => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M4 18H5.5C6.3 18 7 17.3 7 16.5V7.5C7 6.7 6.3 6 5.5 6H4V7H5.5C5.8 7 6 7.2 6 7.5V16.5C6 16.8 5.8 17 5.5 17H4V18ZM13.5 18H10.5C9.7 18 9 17.3 9 16.5V7.5C9 6.7 9.7 6 10.5 6H13.5C14.3 6 15 6.7 15 7.5V16.5C15 17.3 14.3 18 13.5 18ZM10 7.5V16.5C10 16.8 10.2 17 10.5 17H13.5C13.8 17 14 16.8 14 16.5V7.5C14 7.2 13.8 7 13.5 7H10.5C10.2 7 10 7.2 10 7.5ZM20 18H18.5C17.7 18 17 17.3 17 16.5V7.5C17 6.7 17.7 6 18.5 6H20V7H18.5C18.2 7 18 7.2 18 7.5V16.5C18 16.8 18.2 17 18.5 17H20V18Z"
+      fill="black"
+    />
+  </svg>
+);
+
+export default SlidesIconNew;