+++ packages/plugin-gallery/web/src/icons/SlideshowIconNew.tsx
@@ -1,0 +1,24 @@
+/* eslint-disable max-len */
+import React from 'react';
+
+const SlideshowIconNew = props => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="none"
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M16.5 7H7.5C6.7 7 6 7.7 6 8.5V15.5C6 16.3 6.7 17 7.5 17H16.5C17.3 17 18 16.3 18 15.5V8.5C18 7.7 17.3 7 16.5 7ZM17 15.5C17 15.8 16.8 16 16.5 16H7.5C7.2 16 7 15.8 7 15.5V8.5C7 8.2 7.2 8 7.5 8H16.5C16.8 8 17 8.2 17 8.5V15.5ZM3.4 9.59961L1 11.9996L3.4 14.3996L4.1 13.5996L2.4 11.9996L4.1 10.3996L3.4 9.59961ZM19.8999 10.3996L20.5999 9.59961L22.9999 11.9996L20.5999 14.3996L19.8999 13.5996L21.5999 11.9996L19.8999 10.3996Z"
+        fill="black"
+      />
+    </svg>
+  );
+};
+
+export default SlideshowIconNew;