+++ packages/plugin-gallery/web/src/toolbar/galleryLayoutsData.ts
@@ -1,12 +1,21 @@
 import { GALLERY_LAYOUTS } from '../layout-data-provider';
-import { Grid, Masonry, Collage, Thumbnails, Slideshow, Panorama, Slides, Columns } from '../icons';
+import {
+  GridIconNew,
+  MasonryIconNew,
+  CollageIconNew,
+  ThumbnailsIconNew,
+  SlideshowIconNew,
+  PanoramaIconNew,
+  SlidesIconNew,
+  ColumnsIconNew,
+} from '../icons';
 
 export const galleryLayoutsData = [
   {
     text: 'GalleryPlugin_Layout_Grid',
     tooltip: 'GalleryPlugin_Layout_Grid',
     commandKey: GALLERY_LAYOUTS.GRID,
-    icon: Grid,
+    icon: GridIconNew,
     dataHook: 'GalleryPlugin_Layout_Grid',
   },
   {
@@ -14,48 +23,48 @@ export const galleryLayoutsData = [
     tooltip: 'GalleryPlugin_Layout_Masonry',
     commandKey: GALLERY_LAYOUTS.MASONRY,
     dataHook: 'GalleryPlugin_Layout_Masonry',
-    icon: Masonry,
+    icon: MasonryIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.COLLAGE,
     text: 'GalleryPlugin_Layout_Collage',
     tooltip: 'GalleryPlugin_Layout_Collage',
     dataHook: 'GalleryPlugin_Layout_Collage',
-    icon: Collage,
+    icon: CollageIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.THUMBNAIL,
     text: 'GalleryPlugin_Layout_Thumbnails',
     tooltip: 'GalleryPlugin_Layout_Thumbnails',
     dataHook: 'GalleryPlugin_Layout_Thumbnails',
-    icon: Thumbnails,
+    icon: ThumbnailsIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.FULLSIZE,
     text: 'GalleryPlugin_Layout_Slideshow',
     tooltip: 'GalleryPlugin_Layout_Slideshow',
     dataHook: 'GalleryPlugin_Layout_Slideshow',
-    icon: Slideshow,
+    icon: SlideshowIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.PANORAMA,
     text: 'GalleryPlugin_Layout_Panorama',
     tooltip: 'GalleryPlugin_Layout_Panorama',
     dataHook: 'GalleryPlugin_Layout_Panorama',
-    icon: Panorama,
+    icon: PanoramaIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.COLUMN,
     text: 'GalleryPlugin_Layout_Columns',
     tooltip: 'GalleryPlugin_Layout_Columns',
     dataHook: 'GalleryPlugin_Layout_Columns',
-    icon: Columns,
+    icon: ColumnsIconNew,
   },
   {
     commandKey: GALLERY_LAYOUTS.SLIDER,
     text: 'GalleryPlugin_Layout_Slides',
     tooltip: 'GalleryPlugin_Layout_Slides',
     dataHook: 'GalleryPlugin_Layout_Slides',
-    icon: Slides,
+    icon: SlidesIconNew,
   },
 ];