+++ packages/plugin-social-polls/web/src/toolbar/PollButtonsData.ts
@@ -1,17 +1,17 @@
-import { LayoutGridIcon, LayoutListIcon } from '../assets/icons';
+import { LayoutGridIconNew, LayoutListIconNew } from '../assets/icons';
 import CustomizeButton from './CustomizeButton';
 
 export const layoutData = [
   {
     text: 'Poll_PollSettings_Tab_Layout_Section_Answers_Layout_Grid',
     commandKey: 'GRID',
-    icon: LayoutGridIcon,
+    icon: LayoutGridIconNew,
     dataHook: 'poll_grid',
   },
   {
     text: 'Poll_PollSettings_Tab_Layout_Section_Answers_Layout_List',
     commandKey: 'LIST',
-    icon: LayoutListIcon,
+    icon: LayoutListIconNew,
     dataHook: 'poll_list',
   },
   {